+++ tests/Unit/StringTypeTest.php
@@ -126,6 +126,7 @@ class StringTypeTest extends TestCase
 	public function testIfEqualsValueComparesOnlyValues( RepresentsStringType $stringType, RepresentsStringType $anotherStringType, bool $expectedResult ): void
 	{
 		self::assertSame( $expectedResult, $stringType->equalsValue( $anotherStringType ) );
+		self::assertSame( $expectedResult, $stringType->equalsValue( $anotherStringType->toString() ) );
 	}
 
 	public function testIsEmpty(): void
@@ -143,5 +144,15 @@ class StringTypeTest extends TestCase
 		};
 
 		self::assertEquals( 'a', $stringType->toString() );
+	}
+
+	public function testTypeCasting(): void
+	{
+		self::assertEquals( 1.2, (new AnyStringType( '1.2' ))->toFloat() );
+		self::assertEquals( -1.2, (new AnyStringType( '-1.2' ))->toFloat() );
+		self::assertEquals( 0.0, (new AnyStringType( '0.0' ))->toFloat() );
+		self::assertEquals( 120, (new AnyStringType( '120' ))->toInt() );
+		self::assertEquals( -120, (new AnyStringType( '-120' ))->toInt() );
+		self::assertEquals( 0, (new AnyStringType( '0' ))->toInt() );
 	}
 }